Replaced usage of int with size_t in buffer space.

+++ Buffer.h
@@ -16,10 +16,10 @@ class Buffer : public EventTarget
 		uint32_t PacketsRemaining ( );
 
 	protected:
-		Packet* buf;     //!< Pointer to buffer space
-		int buf_index;   //!< Pointer to next empty buffer space
-		int buf_valid;   //!< Pointer to oldest valid data
-		size_t buf_size; //!< Size of buffer
+		Packet* buf;        //!< Pointer to buffer space
+		size_t buf_index;   //!< Pointer to next empty buffer space
+		size_t buf_valid;   //!< Pointer to oldest valid data
+		size_t buf_size;    //!< Size of buffer
 
 	private:
